+++ DemoBots/Components/OrientationComponent.swift
@@ -14,7 +14,7 @@ class OrientationComponent: GKComponent {
     
     var zRotation: CGFloat = 0.0 {
         didSet {
-            let twoPi = CGFloat.pi * 2
+            let twoPi = CGFloat(Double.pi * 2)
             zRotation = (zRotation + twoPi).truncatingRemainder(dividingBy: twoPi)
         }
     }
